Cambios en las variables para manejar mejor los valores

+++ Models/Prestamos.cs
@@ -12,10 +12,10 @@ namespace ReyphillNunez_Ap1_P1.Models
         [StringLength(100)]
 		[RegularExpression("^[a-zA-ZÀ-ÿ\\s]+$", ErrorMessage = "Solo se permiten letras.")]
 		public string? Concepto { get; set; }
-		public decimal Balance { get; set; }
+		public double Balance { get; set; }
 		[Required(ErrorMessage = " Campo obligatorio")]
         [Range(1,double.MaxValue)]
-        public decimal Monto {  get; set; }
+        public double Monto {  get; set; }
         [ForeignKey("Deudor")]
 		[Required(ErrorMessage = "Debe seleccionar un tipo")]
 		public int DeudorId { get; set; }
